update move_a_file api

+++ auto_everything/disk.py
@@ -691,6 +691,8 @@ class Disk:
     def move_a_file(self, source_file_path: str, target_file_path: str):
         source_file_path = self._expand_user(source_file_path)
         target_file_path = self._expand_user(target_file_path)
+        if source_file_path == target_file_path:
+            return
         if self.exists(target_file_path):
             os.remove(target_file_path)
         os.rename(source_file_path, target_file_path)
